fix: [3.0] remove golangci-lint v2 exclusion rules and fix lint violations (#49114)
## Summary
Cherry-pick of #48586 to the 3.0 branch.

- Remove all temporary exclusion rules added during the golangci-lint
v1→v2 upgrade
- Fix ~1500 lint violations (QF1008, ST1005, gosec, revive, ineffassign,
unconvert, depguard, gocritic, staticcheck)

pr: #48586

## Test plan
- [ ] CI passes


---------

+++ internal/datacoord/compaction_task_clustering_test.go
@@ -74,7 +74,7 @@ func (s *ClusteringCompactionTaskSuite) SetupTest() {
 	s.mockAlloc = allocator.NewMockAllocator(s.T())
 	s.mockAlloc.EXPECT().AllocN(mock.Anything).RunAndReturn(func(x int64) (int64, int64, error) {
 		start := s.mockID.Load()
-		end := s.mockID.Add(int64(x))
+		end := s.mockID.Add(x)
 		return start, end, nil
 	}).Maybe()
 	s.mockAlloc.EXPECT().AllocID(mock.Anything).RunAndReturn(func(ctx context.Context) (int64, error) {
